[fix] 커뮤니티 전체조회 퍼센테이지 정수로 리턴

+++ src/main/java/sparta/seed/community/domain/dto/responsedto/CommunityAllResponseDto.java
@@ -9,8 +9,8 @@ public class CommunityAllResponseDto {
 	private String nickname;
 	private String title;
 	private String img;
-	private double currentPercent;
-	private double successPercent;
+	private long currentPercent;
+	private long successPercent;
 	private String dateStatus;
 	private boolean secret;
 	private String password;
@@ -22,8 +22,8 @@ public class CommunityAllResponseDto {
 		this.nickname = nickname;
 		this.title = title;
 		this.img = img;
-		this.currentPercent = currentPercent;
-		this.successPercent = successPercent;
+		this.currentPercent = (long) currentPercent;
+		this.successPercent = (long) successPercent;
 		this.dateStatus = dateStatus;
 		this.secret = secret;
 		this.password = password;
